web-app/api/requirements.txt
Flask==2.3.2
pymongo==4.6.0
python-dotenv==1.0.0